117 admin tools (#119)
* ✨ add api for admin-tools

* 🚧 Start Implementing Admin Page

* 🚧 Progress on Admin Panel

* 🚧 Add Change Email Functionality to the Frontend

* 🚧 Rearranging Admin Modals

* ✨ add mail setup

* 🚧 add some mail stuff + configs

* 🚧 add registrationMode to config, make email config optinal

* 🚧 add isEmailVerified, return on registrationMode, add email functionalities

* 🐛 send emails if emails is enabled

* 🚧 add isUserVerified in userlist

* 🚧 refactor admin services to return the correct data

* 🚧 add name in userentity

* 🚧 disable register only for api registers

* 🚧 Add Admin Create User Frontend Functionality

* 🚧 remove systemuser from list, return user after create, sort userslst

* 🚧 return populated user after creation

* 🚧 return populated user after creation

* 🚧 return populated user after creation

* 🚧 add verifyEmail

* 🚧 Fix Admin Table Sorting

* 🚧 block deleting processing projects, add update name of user,

* 🚧 remove initalUsers

* 🚧 Revise Admin Change User Email Dialog and Logic

* 🚧 add change username dialog

* 🚧 Implement admin initialization logic and effect

* 🚧 add change username dialog

* 🚧  Update email templates for account creation, password reset, and verification

* 🚧 add refresh token action and effect for username change

* 🌐 Add i18n

* 🚧 Start Implementing Frontend Side of Email Verification

* 🚧  prevent email verification for already verified users

* 🚧  prevent email verification for already verified users

* 🚧 add migrations for verified and sizeInBytes, add sizeInBytes to projects

* 🚧 Email Verification Progress

* 🐛  update sticky header position to use variable for consistency

* 🌐  update username references to name for consistency

* 🌐 Add missing i18n

* 🚧 implement logic behind forget passwort

* 🚧  Add contact email configuration and update references in frontend

* 🚧 Add Team schema with properties and validation, not active yet

* ⬆️ angular 20 and cleanups (#118)

* 🌐 Update registration information and remove development warnings

* 💄 Finally Add Revised Logo

* 🚧 Start Migrating to Angular 20

* 🚧 Migrate to Material 20

* ⬆️  Update Dependencies

* 💄 Fix Player Transcript Layout For Short Texts

* 💄  Download Button Better Visible

* 💄 Remove Legal Footer Header If No Link is Available

---------


* 🔥 remove code, fix/check/remove old todos

* 🔥 remove code, fix/check/remove old todos

* ✨ add password forget feature

* 🌐 update i18n

* ✨🚧  add teams to limit user

* 🔧  update config example

* Add user storage info

* 🚧 add isUserSizeLimit in projectCreated, remove unused imports

* 🚧 🌐 cleanups and i18n

* 🌐 i18n

* 🔧 update config & example config

---------

+++ backend/src/resources/project/project.service.ts
@@ -84,6 +84,13 @@ export class ProjectService {
   }
 
   async create(authUser: AuthUser, createProjectDto: CreateProjectDto) {
+    const userlimitReached = await this.permissions.isUserSizeLimitReached(
+      authUser,
+    );
+    if (userlimitReached) {
+      throw new CustomForbiddenException('user_size_limit_reached');
+    }
+
     const status = ProjectStatus.WAITING;
 
     const videosMetadata: UploadMetadata[] = [];
@@ -144,6 +151,7 @@ export class ProjectService {
       status: MediaStatus.WAITING,
       title: 'Main Video',
       resolutions: [],
+      sizeInBytes: 0,
     };
 
     const mainAudio: Audio = {
@@ -153,6 +161,7 @@ export class ProjectService {
       originalFileName: '',
       status: MediaStatus.WAITING,
       title: 'Main Audio',
+      sizeInBytes: 0,
     };
 
     //create project
@@ -477,6 +486,13 @@ export class ProjectService {
       throw new CustomForbiddenException('must_be_owner');
     }
 
+    if (
+      project.status === ProjectStatus.WAITING ||
+      project.status === ProjectStatus.PROCESSING
+    ) {
+      throw new CustomForbiddenException('cannot_delete_project_due_status');
+    }
+
     await Promise.all([
       // Delete project
       this.db.projectModel.findByIdAndDelete(id),
@@ -735,6 +751,7 @@ export class ProjectService {
       status: MediaStatus.WAITING,
       extension: 'mp4',
       resolutions: [],
+      sizeInBytes: 0,
     };
 
     const updatedProject = await this.db.updateProjectByIdAndReturn(projectId, {
